# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/pasien": {
            "get": {
                "description": "Get a list of Pasien.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pasien"
                ],
                "summary": "Get all Pasien.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pasien"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Pasien.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pasien"
                ],
                "summary": "Create New Pasien.",
                "parameters": [
                    {
                        "description": "the body to create a new Pasien",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.pasienInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pasien"
                        }
                    }
                }
            }
        },
        "/pasien/{id}": {
            "get": {
                "description": "Get an Pasien by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pasien"
                ],
                "summary": "Get Pasien.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pasien id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pasien"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Pasien by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pasien"
                ],
                "summary": "Delete one Pasien.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pasien id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Pasien by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pasien"
                ],
                "summary": "Update Pasien.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pasien id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age pasien category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.pasienInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pasien"
                        }
                    }
                }
            }
        },
        "/perawat": {
            "get": {
                "description": "Get a list of Perawat.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Perawat"
                ],
                "summary": "Get all Perawat.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Perawat"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Perawat.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Perawat"
                ],
                "summary": "Create New Perawat.",
                "parameters": [
                    {
                        "description": "the body to create a new Perawat",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.perawatInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Perawat"
                        }
                    }
                }
            }
        },
        "/perawat/{id}": {
            "get": {
                "description": "Get an Perawat by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Perawat"
                ],
                "summary": "Get Perawat.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Perawat id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Perawat"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Perawat by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Perawat"
                ],
                "summary": "Delete one Perawat.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Perawat id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Perawat by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Perawat"
                ],
                "summary": "Update Perawat.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Perawat id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age perawat category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.perawatInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Perawat"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.pasienInput": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_hp": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                }
            }
        },
        "controller.perawatInput": {
            "type": "object",
            "properties": {
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "models.Pasien": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_hp": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Perawat": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
...
    }
}